+++ src/components/calcite-chip/calcite-chip.stories.js
@@ -3,7 +3,7 @@ import { withKnobs, select, boolean } from "@storybook/addon-knobs";
 import {
   darkBackground,
   iconNames,
-  parseReadme
+  parseReadme,
 } from "../../../.storybook/helpers";
 import readme from "./readme.md";
 const notes = parseReadme(readme);
@@ -15,7 +15,13 @@ storiesOf("Chip", module)
     () => `
     <div style="background-color:white;padding:100px">
     <calcite-chip
-    scale="${select("scale", ["xs", "s", "m", "l", "xl"], "m")}"
+    scale="${select("scale", ["s", "m", "l"], "m")}"
+    appearance="${select("appearance", ["solid", "outline"], "solid")}"
+    color="${select(
+      "color",
+      ["blue", "red", "yellow", "green", "grey"],
+      "grey"
+    )}"
     dismissible="${boolean("dismissible", false)}"
     >My great chip</calcite-chip>
     </div>
@@ -28,7 +34,13 @@ storiesOf("Chip", module)
     <div style="background-color:white;padding:100px">
     <calcite-chip
     icon="${select("icon", iconNames, iconNames[0])}"
-    scale="${select("scale", ["xs", "s", "m", "l", "xl"], "m")}"
+    scale="${select("scale", ["s", "m", "l"], "m")}"
+    appearance="${select("appearance", ["solid", "outline"], "solid")}"
+    color="${select(
+      "color",
+      ["blue", "red", "yellow", "green", "grey"],
+      "grey"
+    )}"
     dismissible="${boolean("dismissible", false)}"
     >
     My great chip</calcite-chip>
@@ -41,7 +53,13 @@ storiesOf("Chip", module)
     () => `
     <div style="background-color:white;padding:100px">
     <calcite-chip
-    scale="${select("scale", ["xs", "s", "m", "l", "xl"], "m")}"
+    scale="${select("scale", ["s", "m", "l"], "m")}"
+    appearance="${select("appearance", ["solid", "outline"], "solid")}"
+    color="${select(
+      "color",
+      ["blue", "red", "yellow", "green", "grey"],
+      "grey"
+    )}"
     dismissible="${boolean("dismissible", false)}"
     >
     <img slot="chip-image" src="https://placekitten.com/50/50" />
@@ -56,7 +74,13 @@ storiesOf("Chip", module)
     <div style="background-color:#2b2b2b;padding:100px">
     <calcite-chip
     theme="dark"
-    scale="${select("scale", ["xs", "s", "m", "l", "xl"], "m")}"
+    scale="${select("scale", ["s", "m", "l"], "m")}"
+    appearance="${select("appearance", ["solid", "outline"], "solid")}"
+    color="${select(
+      "color",
+      ["blue", "red", "yellow", "green", "grey"],
+      "grey"
+    )}"
     dismissible="${boolean("dismissible", false)}"
     >My great chip</calcite-chip>
     </div>
@@ -69,7 +93,13 @@ storiesOf("Chip", module)
     <div style="background-color:white;padding:100px" dir="rtl">
     <calcite-chip
     icon="${select("icon", iconNames, iconNames[0])}"
-    scale="${select("scale", ["xs", "s", "m", "l", "xl"], "m")}"
+    scale="${select("scale", ["s", "m", "l"], "m")}"
+    appearance="${select("appearance", ["solid", "outline"], "solid")}"
+    color="${select(
+      "color",
+      ["blue", "red", "yellow", "green", "grey"],
+      "grey"
+    )}"
     dismissible="${boolean("dismissible", false)}"
     >My great chip</calcite-chip>
     </div>