+++ src/components/calcite-chip/calcite-chip.tsx
@@ -27,7 +27,19 @@ export class CalciteChip {
   @Prop({ reflect: true }) theme: "light" | "dark";
 
   /** specify the scale of the chip, defaults to m */
-  @Prop({ reflect: true }) scale: "xs" | "s" | "m" | "l" | "xl" = "m";
+  @Prop({ reflect: true }) scale: "s" | "m" | "l" = "m";
+
+  /** specify the color of the button, defaults to blue */
+  @Prop({ mutable: true, reflect: true }) color:
+    | "blue"
+    | "red"
+    | "yellow"
+    | "green"
+    | "grey" = "grey";
+
+  /** specify the appearance style of the button, defaults to solid. */
+  @Prop({ mutable: true, reflect: true }) appearance: "solid" | "outline" =
+    "solid";
 
   /** optionally pass an icon to display - accepts Calcite UI icon names  */
   @Prop({ reflect: true }) icon?: string;
@@ -69,18 +81,20 @@ export class CalciteChip {
   //
   //--------------------------------------------------------------------------
   connectedCallback() {
-    let scale = ["xs", "s", "m", "l", "xl"];
+    // prop validations
+    let scale = ["s", "m", "l"];
     if (!scale.includes(this.scale)) this.scale = "m";
+
+    let appearance = ["solid", "outline"];
+    if (!appearance.includes(this.appearance)) this.appearance = "solid";
+
+    let color = ["blue", "green", "grey", "yellow", "red"];
+    if (!color.includes(this.color)) this.color = "grey";
   }
 
   render() {
     const dir = getElementDir(this.el);
-    const iconScale =
-      this.scale === "xs" || this.scale === "s" || this.scale === "m"
-        ? "s"
-        : this.scale === "l"
-        ? "m"
-        : "l";
+    const iconScale = this.scale !== "l" ? "s" : "m";
 
     const iconEl = (
       <calcite-icon
@@ -102,8 +116,8 @@ export class CalciteChip {
 
     return (
       <Host dir={dir}>
-        {this.icon ? iconEl : null}
         <slot name="chip-image"></slot>
+        {this.icon ? iconEl : null}
         <span>
           <slot />
         </span>