+++ CHANGELOG.md
@@ -11,12 +11,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `calcite-checkbox` - `size` prop is now `scale` to be consistent with other components
 - `calcite-chip` - will not show the dismiss ("x") button unless new `dismissible` prop is `true`
+- `calcite-chip` - will no longer accept `xs` or `xl` values for `scale`
+- `calcite-combobox` - will no longer accept `xs` or `xl` values for `scale`
 
 ### Added
 
 - `calcite-radio-group-item` can now display an icon by passing a Calcite UI Icon name to the `icon` attribute. The icon can be positioned with the `icon-position` attribute.
 - `calcite-split-button` now accepts `ellipsis` as a value for the `dropdown-icon-type` attribute
 - `calcite-graph` component for simple area graphs from series data sets
+- `calcite-chip` - now has a `color` prop that will accept `grey` (default), `blue`, `red`, `yellow`, and `green` as values
+- `calcite-chip` - now has an `appearance` prop that will accept `solid` (default), `outline`, and `clear` as values
 
 ### Fixed